First files
In these files I am hard coding the start of the timeline to originate at the earth's formation. I am using an array and passing through the Year, month and day as a string. This String is then converted into a float, I intend on turning these floats into seconds from the earth's creation then using a sorting algorithm to sort it from the beginning (earth's creation) to the end (end of the universe)