DateOrdering.py produced
This was done in conjunction with OpenAI's ChatGPT. This will allow me to create an ordered lists of dates. You first enter the dates, and it uses Pythons inbuilt datetime library it will order the dates for you. These dates are outputted as a list in the console. If an invalid date is entered, then a value error will occur, and it will prompt the user to reenter the date.